--- packages/babel-plugin-core-web/detectors/Blob.js
@@ -1,9 +1,0 @@
-const m = require("../lib/ast-matcher");
-const p = require("path");
-
-module.exports = [
-	m("Blob"),
-	m("BlobBuilder"),
-	m("URL.createObjectURL"),
-	m("URL.revokeObjectURL")
-];
