--- packages/babel-plugin-core-web/detectors/Element.js
@@ -1,8 +1,0 @@
-const m = require("../lib/ast-matcher");
-
-module.exports = [
-	m("Element"),
-	m("HTMLElement"),
-	m("hasAttribute"),
-	m("document.createElement")
-];
