--- packages/babel-plugin-core-web/detectors/HTMLTemplateElement.js
@@ -1,6 +1,0 @@
-const m = require("../lib/ast-matcher");
-
-module.exports = [
-	m("HTMLTemplateElement"),
-	m(`document.createElement('template')`)
-];
