--- packages/babel-plugin-core-web/detectors/IntersectionObserverEntry.js
@@ -1,3 +1,0 @@
-const m = require("../lib/ast-matcher");
-
-module.exports = [m("isIntersecting"), m("IntersectionObserverEntry")];
