--- packages/babel-plugin-core-web/detectors/Intl.js
@@ -1,9 +1,0 @@
-const m = require("../lib/ast-matcher");
-
-module.exports = [
-	m("Intl"),
-	m("localeCompare"),
-	m("toLocaleString"),
-	m("toLocaleDateString"),
-	m("toLocaleTimeString")
-];
