--- packages/babel-plugin-core-web/detectors/URL.js
@@ -1,3 +1,0 @@
-const m = require("../lib/ast-matcher");
-
-module.exports = [m("URL"), m("URLSearchParams")];
