--- packages/babel-plugin-core-web/detectors/WebAnimations.js
@@ -1,3 +1,0 @@
-const m = require("../lib/ast-matcher");
-
-module.exports = [m("animate"), m("Animation")];
