--- packages/babel-plugin-core-web/detectors/atob.js
@@ -1,4 +1,0 @@
-const m = require("../lib/ast-matcher");
-const p = require("path");
-
-module.exports = [m("btoa"), m("atob")];
