--- packages/babel-plugin-core-web/detectors/fetch.js
@@ -1,9 +1,0 @@
-const m = require("../lib/ast-matcher");
-
-module.exports = [
-	m("fetch"),
-	m("Headers"),
-	m("Body"),
-	m("Request"),
-	m("Response")
-];
