--- packages/babel-plugin-core-web/detectors/matchMedia.js
@@ -1,3 +1,0 @@
-const m = require("../lib/ast-matcher");
-
-module.exports = [m("matchMedia"), m("MediaQueryList")];
