--- packages/babel-plugin-core-web/detectors/~custom-elements.js
@@ -1,6 +1,0 @@
-const m = require("../lib/ast-matcher");
-
-module.exports = [
-	m("customElements.define"),
-	m("customElements.polyfillWrapFlushCallback")
-];
