--- packages/babel-plugin-core-web/detectors/~shadycss-custom-style-interface.js
@@ -1,5 +1,0 @@
-const m = require("../lib/ast-matcher");
-
-module.exports = [
-	// m('CustomStyleInterface'),
-];
