--- packages/babel-plugin-core-web/detectors/~shadydom.js
@@ -1,3 +1,0 @@
-const m = require("../lib/ast-matcher");
-
-module.exports = [m("ShadyDOM"), m("shadowRoot"), m("attachShadow")];
