--- packages/babel-plugin-core-web/detectors/~viewport.js
@@ -1,10 +1,0 @@
-const m = require("../lib/ast-matcher");
-
-module.exports = [
-	m("innerHeight"),
-	m("innerWidth"),
-	m("pageXOffset"),
-	m("pageYOffset"),
-	m("scrollX"),
-	m("scrollY")
-];
