+++ packages/babel-plugin-core-web/matchers/Element.js
@@ -1,0 +1,8 @@
+const m = require("../lib/ast-matcher");
+
+module.exports = [
+	m("Element"),
+	m("HTMLElement"),
+	m("hasAttribute"),
+	m("document.createElement")
+];
