+++ packages/babel-plugin-core-web/matchers/HTMLTemplateElement.js
@@ -1,0 +1,6 @@
+const m = require("../lib/ast-matcher");
+
+module.exports = [
+	m("HTMLTemplateElement"),
+	m(`document.createElement('template')`)
+];
