+++ packages/babel-plugin-core-web/matchers/Intl.js
@@ -1,0 +1,9 @@
+const m = require("../lib/ast-matcher");
+
+module.exports = [
+	m("Intl"),
+	m("localeCompare"),
+	m("toLocaleString"),
+	m("toLocaleDateString"),
+	m("toLocaleTimeString")
+];
