+++ packages/babel-plugin-core-web/matchers/WebAnimations.js
@@ -1,0 +1,3 @@
+const m = require("../lib/ast-matcher");
+
+module.exports = [m("animate"), m("Animation")];
