+++ packages/babel-plugin-core-web/matchers/atob.js
@@ -1,0 +1,4 @@
+const m = require("../lib/ast-matcher");
+const p = require("path");
+
+module.exports = [m("btoa"), m("atob")];
