+++ packages/babel-plugin-core-web/matchers/fetch.js
@@ -1,0 +1,9 @@
+const m = require("../lib/ast-matcher");
+
+module.exports = [
+	m("fetch"),
+	m("Headers"),
+	m("Body"),
+	m("Request"),
+	m("Response")
+];
