+++ packages/babel-plugin-core-web/matchers/~custom-elements.js
@@ -1,0 +1,6 @@
+const m = require("../lib/ast-matcher");
+
+module.exports = [
+	m("customElements.define"),
+	m("customElements.polyfillWrapFlushCallback")
+];
