+++ packages/babel-plugin-core-web/matchers/~shadycss-apply-shim.js
@@ -1,0 +1,3 @@
+const m = require("../lib/ast-matcher");
+
+module.exports = [m("ShadyCSS")];
