+++ packages/babel-plugin-core-web/matchers/~shadycss-custom-style-interface.js
@@ -1,0 +1,5 @@
+const m = require("../lib/ast-matcher");
+
+module.exports = [
+	// m('CustomStyleInterface'),
+];
