+++ packages/babel-plugin-core-web/matchers/~shadydom.js
@@ -1,0 +1,3 @@
+const m = require("../lib/ast-matcher");
+
+module.exports = [m("ShadyDOM"), m("shadowRoot"), m("attachShadow")];
