+++ packages/babel-plugin-core-web/matchers/~viewport.js
@@ -1,0 +1,10 @@
+const m = require("../lib/ast-matcher");
+
+module.exports = [
+	m("innerHeight"),
+	m("innerWidth"),
+	m("pageXOffset"),
+	m("pageYOffset"),
+	m("scrollX"),
+	m("scrollY")
+];
