+++ packages/babel-plugin-core-web/package.json
@@ -20,7 +20,7 @@
   ],
   "files": [
     "/lib",
-    "/detectors",
+    "/matchers",
     "/LICENSE"
   ],
   "dependencies": {
