+++ packages/core-web-generator/package.json
@@ -3,7 +3,8 @@
   "version": "0.4.0",
   "private": true,
   "scripts": {
-    "build": "node ./generate.js",
+    "build": "tsc && node ./dist/generate.js",
+    "tsc": "tsc",
     "test": "node ./__tests__/core-web-generator.test.js"
   },
   "directories": {
@@ -15,12 +16,12 @@
     "@webcomponents/shadycss": "^1.10.2",
     "@webcomponents/shadydom": "^1.8.0",
     "@webcomponents/template": "^1.4.4",
-    "denodeify": "^1.2.1",
-    "fast-deep-equal": "^3.1.3",
-    "from2-string": "^1.1.0",
-    "merge2": "^1.4.1",
     "polyfill-library": "^3.100.0",
-    "semver": "^7.3.4",
-    "stream-to-string": "^1.2.0"
+    "semver": "^7.3.4"
+  },
+  "devDependencies": {
+    "@types/node": "^14.14.16",
+    "@types/node": "^14.14.16",
+    "typescript": "^4.1.3"
   }
 }
