+++ packages/core-web-generator/src/global.d.ts
@@ -1,0 +1,5 @@
+declare module 'polyfill-library';
+declare module 'stream-to-string';
+declare module 'polyfill-library/lib/sources.js';
+declare module 'path';
+declare module 'semver';
